requirements.txt
absl-py==0.10.0
astor==0.8.1
astropy==4.0.1.post1
astunparse==1.6.3
async-generator==1.10
atomicwrites==1.4.0
backcall==0.2.0
beautifulsoup4==4.9.3
brotlipy==0.7.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cmarkgfm==0.4.2
colorama==0.4.3
cryptography
cycler==0.10.0
cytoolz==0.11.0 
decorator==4.4.2
defusedxml==0.6.0
docutils==0.16
entrypoints==0.3
Flask==1.1.2
future==0.18.2
gast==0.2.2
google-auth==1.21.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
html5lib==1.1
importlib-metadata==1.7.0
iniconfig==1.0.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
Keras==2.1.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
maskgalaxylib==0.0.1
matplotlib==3.3.2
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
numpy==1.18.5
oauthlib==3.1.0
olefile==0.46
opencv-python==4.4.0.42
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.1
pandocfilters==1.4.2
parso==0.7.1
Pillow==7.2.0
pkginfo==1.5.0.1
pluggy==0.13.1
prometheus-client==0.8.0
prompt-toolkit==3.0.7
protobuf==3.13.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pyparsing==2.4.7
pytest==6.1.1
pytest-runner==5.2
python-dateutil==2.8.1
pytz==2020.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.2
readme-renderer==24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rsa==4.6
scikit-image==0.17.2
scipy==1.4.1
Send2Trash==1.5.0
six==1.15.0
soupsieve==2.0.1
tensorboard==1.15.0
tensorboard-plugin-wit==1.7.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
terminado==0.9.1
testpath==0.4.4
tifffile==2020.9.3
toml==0.10.1
tornado==6.0.4
traitlets==5.0.3
webencodings==0.5.1
Werkzeug==1.0.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.12.1
zipp==3.1.0